feat: move reindex action to DocumentUserActions component
